+++ CHANGELOG.md
@@ -81,6 +81,7 @@ that can cause contrast issues.
   - the autocomplete input is now fully disabled when the select is disabled
   - autocomplete filter is now cleared when the select value is set to a falsy value
   - focus last selected option on select opening after setting value in JS
+  - handle async option loading with an initial value (the label is now refreshed when the options are loaded)
 
 ### Security